refactor: extract shared trigger loop methods into TriggerLoopHelpers concern
Extract find_last_agent_comment, update_loop_data, and post_system_notice
from TriggerLoopCheckJob and TriggerLoopVerifyJob into a shared concern
to eliminate code duplication.
